Finish chart data reducer.
