+++ app/middleware/receiveChartType.js
@@ -7,7 +7,6 @@ import {
   RECEIVE_CHART_OPTIONS,
   RECEIVE_DEFAULTS_APPLIED_TO,
 } from '../constants';
-import dispatchChartData from './utils/dispatchChartData';
 import applyChartTypeDefaults from './utils/applyChartTypeDefaults';
 import applyTickFormatters from './utils/applyTickFormatters';
 import applyYDomain from './utils/applyYDomain';
@@ -19,31 +18,6 @@ export default function receiveChartType({ getState }) {
       return dispatch(action);
     }
     const nextConfig = action.data.config;
-
-    /**
-     * Dispatch chart data if chartData not already set up
-     * or if dataFormat has changed
-     */
-    function _shouldDispatchChartData() {
-      return !getState().chartData.length || // chartData not already set up
-        !getState().chartType.config || //  chartType not already set up
-        !getState().chartType.config.dataFormat || // chartType didn't have dataFormat
-        getState().chartType.config.dataFormat !== nextConfig.dataFormat; // dataFormat has changed
-    }
-
-    /**
-     * Send chartData to store if dataFormat has changed
-     */
-    if (_shouldDispatchChartData()) {
-      console.log('from chart type');
-      dispatchChartData(
-        dispatch,
-        nextConfig,
-        getState().transformedData,
-        getState().chartOptions.color,
-        action.src
-      );
-    }
 
     /**
      * Setup chart type default options if NOT bootstrapping from postMessage
