--- app/middleware/transformData.js
@@ -1,24 +1,0 @@
-/**
- * TRANSFORM_DATA middleware. If we have a chart type config
- * with dataFormat, dispatch chartData to store
- */
-
-import { TRANSFORM_DATA } from '../constants';
-import dispatchChartData from './utils/dispatchChartData';
-
-export default function setChartData({ getState }) {
-  return (dispatch) => (action) => {
-    if (TRANSFORM_DATA === action.type) {
-      // will validate input and dispatch formatted data to store
-      console.log('from transform');
-      dispatchChartData(
-        dispatch,
-        getState().chartType.config || {},
-        action.data,
-        getState().chartOptions.color,
-        action.src
-      );
-    }
-    return dispatch(action);
-  };
-}
