+++ app/reducers/genericReducers.js
@@ -36,13 +36,3 @@ export function mergeReducer(initState, allowActions) {
     return state;
   };
 }
-
-export function createReducer(initialState, actionMap) {
-  return (state = initialState, action) => {
-    if (actionMap[action.type]) {
-      return actionMap[action.type](state, action);
-    }
-
-    return state;
-  };
-}
